Exception handling

+++ src/main/java/com/smartprocess/sp_commerce/services/ProductService.java
@@ -2,14 +2,21 @@
 package com.smartprocess.sp_commerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.smartprocess.sp_commerce.dto.ProductDto;
 import com.smartprocess.sp_commerce.entities.Product;
 import com.smartprocess.sp_commerce.repositories.ProductRepository;
+import com.smartprocess.sp_commerce.services.exceptions.DataBaseException;
+import com.smartprocess.sp_commerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -19,7 +26,8 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDto findById(Long id) {
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDto(product);
 	}
 	
@@ -39,15 +47,31 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDto update(Long id, ProductDto dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDto(entity);
+		
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDto(entity);
+			
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException ("Recurso não encontrado");
+		}
+		
+		
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DataBaseException("Falha de integridade referencial");
+	   	}
 	}
